Remove unnecessary conditional
`PublicBody.extract_domain_from_email` returns a `String` domain if one
can be extracted, or `nil` if there's no match. This means we can just
pass `from_email` directly to the method irrespective of its value, and
just ensure we convert whatever result we get to a `String`.

    PublicBody.extract_domain_from_email(nil)
    # => nil
    PublicBody.extract_domain_from_email('foo')
    # => nil
    PublicBody.extract_domain_from_email('foo@example.com')
    # => "example.com"

In the event that `from_email` is `nil`, or there's no match, calling
`#to_s` on the `nil` return value performs the same function as the
`else` clause of the previous conditional.